bump murmur3 to v1.1.1
murmur3 included unsafe code; go1.14 has a -d=checkptr flag that checks
some uses of unsafe code. Turns out, the original usage of murmur3 was
unsafe due to alignment issues.

This does not affect the murmur128 assembly code because unaligned reads
on amd64 are safe (although slower, but we don't usually expect the hash
to have unaligned input).

This bumps the dep to pull in the fixes.